pyproject.toml
[project]
name = "agentic-examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "conventional-pre-commit>=4.4.0",
    "langchain>=1.2.15",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.2.1",
    "langgraph>=1.1.4",
    "pre-commit>=4.5.1",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.2",
]

[tool.ruff]
line-length = 79
lint.select = [
    "ANN",
    "S",
    "BLE",
    "FBT",
    "B",
    "A",
    "COM",
    "C",
    "DTZ",
    "T10",
    "EM",
    "EXE",
    "ISC",
    "ICN",
    "LOG",
    "G",
    "PIE",
    "T20",
    "PYI",
    "PT",
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SIM",
    "TID",
    "TC",
    "ARG",
    "PTH",
    "I",
    "C90",
    "NPY",
    "PD",
    "N",
    "PERF",
    "E",
    "W",
    "D",
    "F",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "FURB",
    "RUF",
    "TRY",
]

lint.ignore = [
    "D212",   # Conflicted with D213
    "D203",   # Conflicted with D211
    "COM812", # To avoid formatting potential issues
    "T201",   # To allow printing
    "BLE001", # To allow blind exceptions for simplicity
]
